More cleanup. Remove openSSL entirely from project

+++ include/Partitioning/CoordinatedPartitionState.hpp
@@ -110,19 +110,19 @@ std::shared_ptr<Record<T>> CoordinatedPartitionState<T>::getRecord(
 template <typename T>
 int CoordinatedPartitionState<T>::getMachineLoad(const int m) const {
   std::lock_guard<std::mutex> lock(*machines_load_edges_mutex);
-  return machines_load_edges.at(m);
+  return (int)machines_load_edges.at(m);
 }
 
 template <typename T>
 int CoordinatedPartitionState<T>::getMachineWeight(const int m) const {
   std::lock_guard<std::mutex> lock(*machines_weight_edges_mutex);
-  return machines_weight_edges.at(m);
+  return (int)machines_weight_edges.at(m);
 }
 
 template <typename T>
 int CoordinatedPartitionState<T>::getMachineLoadVertices(const int m) const {
   std::lock_guard<std::mutex> lock(*machines_load_vertices_mutex);
-  return machines_load_vertices.at(m);
+  return (int)machines_load_vertices.at(m);
 }
 template <typename T>
 void CoordinatedPartitionState<T>::incrementMachineLoad(const int m,
